requirements.txt
bs4==4.12.3
docx==1.1.2
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.18.0
folium==0.17.0
fuzzywuzzy==0.18.0
geopy==2.4.1
markdown==3.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
openpyxl==3.1.5
pandas==2.2.2
pdfplumber==0.11.4
Pillow==10.4.0
pyperclip==1.9.0
python-docx==1.1.2
pywin32; platform_system == "Windows"
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
st_copy_to_clipboard==0.1.6
streamlit_folium==0.22.0
streamlit_js_eval==0.1.7
streamlit_lottie==0.0.5
streamlit==1.37.1
streamlit-option-menu==0.3.13